# docs/swagger.yaml
definitions:
  github_com_andruwizz_gin-rest-book-backend_internal_delivery_request.BookCreate:
    properties:
      author:
        type: string
      title:
        type: string
    required:
    - author
    - title
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_delivery_request.BookUpdate:
    properties:
      author:
        type: string
      title:
        type: string
    required:
    - author
    - title
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_delivery_request.UserCreate:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_delivery_request.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.DataResponse-entity_AuthToken:
    properties:
      data:
        $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_entity.AuthToken'
      success:
        type: boolean
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.DataResponse-entity_Book:
    properties:
      data:
        $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_entity.Book'
      success:
        type: boolean
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.DataResponse-entity_User:
    properties:
      data:
        $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_entity.User'
      success:
        type: boolean
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.EmptyResponse:
    properties:
      success:
        type: boolean
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.errorDetail'
      success:
        type: boolean
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.ListMeta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.ListResponse-entity_Book:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_entity.Book'
        type: array
      meta:
        $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.ListMeta'
      success:
        type: boolean
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.errorDetail:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_entity.AuthToken:
    properties:
      token:
        type: string
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_entity.Book:
    properties:
      author:
        type: string
      created_at:
        type: integer
      id:
        type: string
      title:
        type: string
      updated_at:
        type: integer
    type: object
  github_com_andruwizz_gin-rest-book-backend_internal_entity.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /books:
    get:
      consumes:
      - application/json
      description: Get list of books data
      parameters:
      - description: Item count per page
        in: query
        name: limit
        type: integer
      - description: List page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.ListResponse-entity_Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.ErrorResponse'
      summary: List all book
      tags:
      - Book
    post:
      consumes:
      - application/json
      description: Input new book with title and author name
      parameters:
      - description: Request body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_request.BookCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.DataResponse-entity_Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.ErrorResponse'
      summary: Create a new book
      tags:
      - Book
  /books/{id}:
    delete:
      consumes:
      - application/json
      description: Delete single book data
      parameters:
      - description: Book Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.ErrorResponse'
      summary: Delete a book
      tags:
      - Book
    get:
      consumes:
      - application/json
      description: Get single book data
      parameters:
      - description: Book Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.DataResponse-entity_Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.ErrorResponse'
      summary: Find a book
      tags:
      - Book
    put:
      consumes:
      - application/json
      description: Update existing book
      parameters:
      - description: Book Id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_request.BookUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.DataResponse-entity_Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.ErrorResponse'
      summary: Update a book
      tags:
      - Book
  /users/current:
    get:
      consumes:
      - application/json
      description: Show current authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.DataResponse-entity_User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get current user
      tags:
      - User
  /users/login:
    post:
      consumes:
      - application/json
      description: Login as registered user
      parameters:
      - description: Request body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_request.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.DataResponse-entity_AuthToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.ErrorResponse'
      summary: Login as user
      tags:
      - User
  /users/register:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: Request body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_request.UserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.DataResponse-entity_User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_andruwizz_gin-rest-book-backend_internal_delivery_response.ErrorResponse'
      summary: Register a new user
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"